# Automatically read and write bytecode