--- a/NEWS
+++ b/NEWS
@@ -4,6 +4,10 @@ Changes from 0.12.1
    * `let` has been removed. Python's scoping rules do not make a proper
      implementation of it possible. Use `setv` instead.
    * `lambda` has been removed, but `fn` now does exactly what `lambda` did
+   * Importing or executing a Hy file automatically byte-compiles it, or loads
+     a byte-compiled version if it exists and is up to date. This brings big
+     speed boosts, even for one-liners, because Hy no longer needs to recompile
+     its standard library for every startup.
    * Added bytestring literals, which create `bytes` objects under Python 3
      and `str` objects under Python 2
    * Commas and underscores are allowed in numeric literals