--- a/setup.cfg
+++ b/setup.cfg
@@ -6,3 +6,16 @@ nocapture=1
 
 [wheel]
 universal = 1
+
+[coverage:run]
+omit =
+    */python?.?/*
+    */lib-python/?.?/*.py
+    */lib_pypy/_*.py
+    */site-packages/nose/*
+    */pypy/*
+
+[coverage:report]
+exclude_lines =
+    # Have to re-enable the standard pragma
+    pragma: no cover