--- a/setup.cfg
+++ b/setup.cfg
@@ -19,3 +19,6 @@ omit =
 exclude_lines =
     # Have to re-enable the standard pragma
     pragma: no cover
+# We want ignore_errors so we don't get NoSource warnings for loading
+# byte-compiled Hy modules.
+ignore_errors = True