--- a/tests/importer/test_importer.py
+++ b/tests/importer/test_importer.py
@@ -1,7 +1,9 @@
-from hy.importer import import_file_to_module, import_buffer_to_ast, MetaLoader
+from hy.importer import (import_file_to_module, import_buffer_to_ast,
+                         MetaLoader, get_bytecode_path)
 from hy.errors import HyTypeError
 import os
 import ast
+import tempfile
 
 
 def test_basics():
@@ -11,7 +13,6 @@ def test_basics():
 
 
 def test_stringer():
-    "Make sure the basics of the importer work"
     _ast = import_buffer_to_ast("(defn square [x] (* x x))", '')
     assert type(_ast.body[0]) == ast.FunctionDef
 
@@ -41,3 +42,21 @@ def test_import_error_reporting():
 
     assert _import_error_test() == "Error reported"
     assert _import_error_test() is not None
+
+
+def test_import_autocompiles():
+    "Test that (import) byte-compiles the module."
+
+    f = tempfile.NamedTemporaryFile(suffix='.hy', delete=False)
+    f.write(b'(defn pyctest [s] (+ "X" s "Y"))')
+    f.close()
+
+    try:
+        os.remove(get_bytecode_path(f.name))
+    except (IOError, OSError):
+        pass
+    import_file_to_module("mymodule", f.name)
+    assert os.path.exists(get_bytecode_path(f.name))
+
+    os.remove(f.name)
+    os.remove(get_bytecode_path(f.name))