--- a/tests/resources/bin/bytecompile.hy
+++ b/tests/resources/bin/bytecompile.hy
@@ -1,0 +1,5 @@
+(defmacro m []
+  (print "Hello from macro")
+  "boink")
+
+(print "The macro returned:" (m))