@Kodiologist 2017-04-08
Dangit, still failing on CPython 3.3 and Pypy 2. I'll take a look at them later.