@Kodiologist 2017-04-09
Well, there's not much I can do about Travis's complaint, since I was able to coax PyPy into no longer creating that error for [a nonexistent test](https://github.com/nose-devs/nose/issues/233) on my own machine, with Ubuntu 16.10, but Travis's Python VM runs Ubuntu 12.04. (As has previously been noted, we need to move away from the dated Nose.) I have no real way to debug this. Unless anybody happens to have a machine running Ubuntu 12.04 sitting around, I'll have to disable testing PyPy compilation on Travis, I guess by checking the environment variable `TRAVIS`.