@kirbyfan64 2017-04-09
@Kodiologist I've been able to reproduce the same failures using PyPy 2.5.0 (binaries are [here](https://bitbucket.org/pypy/pypy/downloads/pypy-2.5.0-linux64.tar.bz2)). No Ubuntu 12 needed!