@kirbyfan64 2017-04-09
(FWIW, I haven't reviewed this yet, but so far I'm super glad this has happened; IME the main slowdowns with Hy were always due to the compilation times anyway.)