@Kodiologist 2017-04-09
> @Kodiologist I've been able to reproduce the same failures using PyPy 2.5.0 (binaries are here).

Thanks. I see that these binaries don't have the PEP 3147 backport. Perhaps the reason for the Travis failures is that Ubuntu 12.04 (and possibly also Debian at the time) hadn't yet included the backport. All this suggests that I should be setting `DEBIAN_PYPY_2` only if we're using a sufficiently recent build of PyPy. I have `/usr/share/doc/pypy/changelog.Debian.gz` on my machine, but it doesn't seem to indicate at what version the backport was added. So I have no idea what check I should use.