@Kodiologist 2017-04-10
It turns out although Python 3.3 and the backported PyPy don't have `importlib.util.cache_from_source`, they do have `imp.cache_from_source`. That should make things a lot easier.