@kirbyfan64 2017-04-10
@Kodiologist Hmm...I think maybe I can try it on a Ubuntu Docker image...Let's see...