@Kodiologist 2017-04-10
I was able to install the Ubuntu 12.04 package for PyPy on my own system, but I can't create a virtualenv because of a crash from, suitably enough, `imp.cache_from_source` (`ValueError: Not a .py file: /tmp/pip-build-zd1NUk/six/six-1.10.0.dist-info/DESCRIPTION.rst`).