@Kodiologist 2017-04-10
Finally passing. To summarize what I learned, Travis does *not* use Debian PyPy. It uses a specific PyPy nightly, which I was able to get from [here](http://buildbot.pypy.org/nightly/release-2.5.x/). I could then replicate the bug on my own machine, but I didn't find a way to really fix it. Instead, I've just changed `hy.importer` not to compile `__init.hy__` files when running PyPy under Nose. (It works fine when you don't use Nose.)